+++ src/main/factories/controllers/login/login-controller-factory.ts
@@ -1,4 +1,4 @@
-import { LoginController } from '../../../../presentation/controllers/login/login-controller'
+import { LoginController } from '../../../../presentation/controllers/login/login/login-controller'
 import { Controller } from '../../../../presentation/protocols'
 import { makeLogControllerDecorator } from '../../decorators/log-controller-decorator-factory'
 import { makeDbAuthentication } from '../../usecases/authentication/db-authentication-factory'
