--- src/presentation/controllers/login/login-controller-protocols.ts
@@ -1,2 +1,0 @@
-export * from '../../../domain/usecases/authentication'
-export * from '../../protocols'
