+++ src/presentation/controllers/login/login/login-controller.spec.ts
@@ -1,6 +1,6 @@
 import { Authentication, AuthenticationModel, HttpRequest, Validation } from './login-controller-protocols'
-import { MissingParamError } from '../../errors'
-import { badRequest, ok, serverError, unauthorized } from '../../helpers/http/http-helper'
+import { MissingParamError } from '../../../errors'
+import { badRequest, ok, serverError, unauthorized } from '../../../helpers/http/http-helper'
 import { LoginController } from './login-controller'
 
 const makeAuthentication = (): Authentication => {
